+++ providers/windows/boottime_windows_test.go
@@ -18,26 +18,16 @@
 package windows
 
 import (
-	"time"
+	"testing"
 
-	"github.com/pkg/errors"
-
-	windows "github.com/elastic/go-windows"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func BootTime() (time.Time, error) {
-	msSinceBoot, err := windows.GetTickCount64()
-	if err != nil {
-		return time.Time{}, errors.Wrap(err, "failed to get boot time")
-	}
+func TestBootTime(t *testing.T) {
+	bootTime, err := BootTime()
+	require.NoError(t, err)
 
-	bootTime := time.Now().Add(-1 * time.Duration(msSinceBoot) * time.Millisecond)
-
-	// According to GetTickCount64, the resolution of the value is limited to
-	// the resolution of the system timer, which is typically in the range of
-	// 10 milliseconds to 16 milliseconds. So this will round the value to the
-	// nearest second to not mislead anyone about the precision of the value
-	// and to provide a stable value.
-	bootTime = bootTime.Round(time.Second)
-	return bootTime, nil
+	// There should be no sub-second precision in the time.
+	assert.Equal(t, 0, bootTime.Nanosecond())
 }