+++ system_test.go
@@ -22,8 +22,8 @@ import (
 	"os"
 	osUser "os/user"
 	"runtime"
+	"sort"
 	"strconv"
-	"strings"
 	"syscall"
 	"testing"
 	"time"
@@ -162,20 +162,23 @@ func TestSelf(t *testing.T) {
 	}
 
 	if v, ok := process.(types.Environment); ok {
-		expectedEnv := map[string]string{}
-		for _, keyValue := range os.Environ() {
-			parts := strings.SplitN(keyValue, "=", 2)
-			if len(parts) != 2 {
-				t.Fatal("failed to parse os.Environ()")
-			}
-			expectedEnv[parts[0]] = parts[1]
-		}
 		actualEnv, err := v.Environment()
 		if err != nil {
 			t.Fatal(err)
 		}
-		assert.Equal(t, expectedEnv, actualEnv)
 		output["process.env"] = actualEnv
+
+		// Format the output to match format from os.Environ().
+		keyEqualsValueList := make([]string, 0, len(actualEnv))
+		for k, v := range actualEnv {
+			keyEqualsValueList = append(keyEqualsValueList, k+"="+v)
+		}
+		sort.Strings(keyEqualsValueList)
+
+		expectedEnv := os.Environ()
+		sort.Strings(expectedEnv)
+
+		assert.Equal(t, expectedEnv, keyEqualsValueList)
 	}
 
 	memInfo, err := process.Memory()